pyproject.toml
[project]
name = "gradebooks-takehome"
version = "0.1.0"
description = ""
authors = [
    {name = "lihaojin",email = "haojin.li14@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.8,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
